refactor(invoicing): split the draft lifecycle out of attendee_invoice_service (#911)
CI file-length gate: attendee_invoice_service.py reached 1058 lines against the
1000 cap. It was 880 before this PR, so the breach is this branch's doing.

Move the "Phase 5: Editing, issuing, deletion" section into
events/service/attendee_invoice_draft_service.py — everything acting on an
invoice that already exists, where generation, PDF rendering and delivery stay
next door. Parent drops to 710, new module is 316.

The seam was chosen for coupling: the section needs exactly one thing from the
parent, _generate_and_save_pdf, which stays put so the
attendee_invoice_service.render_pdf patch target every test uses is unchanged.
The credit-note section would have been the more natural domain boundary but
depends on six private helpers there.

Also: SubmittedLineItem and DerivedTotals move to models/attendee_invoice.py
alongside BuyerBillingSnapshot, InvoiceLineItemDict and InvoiceVatBucketDict,
which is where this model's typed data shapes already live. The credit-note
path's delete_draft_invoice call becomes a function-local import to avoid a
cycle, matching issue_and_deliver's existing task import.

Comments trimmed where this branch had restated the same #911 rationale in
several places; the canonical version now lives on DERIVED_TOTAL_FIELDS.
