+++ src/events/service/attendee_invoice_service.py
@@ -11,7 +11,6 @@ import structlog
 from django.core.files.base import ContentFile
 from django.db import transaction
 from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
 from ninja.errors import HttpError
 
 from common.constants import EU_MEMBER_STATES
@@ -343,358 +342,6 @@ def ensure_pdf_exists(invoice: AttendeeInvoice) -> None:
 
 
 # ---------------------------------------------------------------------------
-# Phase 5: Editing, issuing, deletion
-# ---------------------------------------------------------------------------
-
-
-EDITABLE_DRAFT_FIELDS = frozenset(
-    {
-        "buyer_name",
-        "buyer_vat_id",
-        "buyer_vat_country",
-        "buyer_address",
-        "buyer_email",
-        "currency",
-        "reverse_charge",
-        "discount_code_text",
-        "line_items",
-    }
-)
-
-# Money columns derived from ``line_items``, never independently writable (#911).
-# They used to sit in EDITABLE_DRAFT_FIELDS, so a PATCH of ``total_gross`` alone
-# left a header claiming EUR 1000 above line items still summing to EUR 111 -- a
-# contradiction that shipped in one response once ``vat_breakdown`` was exposed
-# (#910), and that ``issue_draft_invoice`` then turned into a delivered PDF.
-# ``update_draft_invoice`` recomputes them whenever the lines change instead, so
-# the header cannot disagree with the rows it summarizes.
-DERIVED_TOTAL_FIELDS = ("total_gross", "total_net", "total_vat", "vat_rate", "discount_amount_total")
-
-
-class SubmittedLineItem(t.TypedDict, total=False):
-    """A line item as submitted for an edit, before normalization.
-
-    ``InvoiceLineItemSchema`` parses the amounts as ``Decimal``; the JSON column
-    stores strings. ``total=False`` because :func:`_normalize_line_items` defaults
-    every key -- completing a partial item is precisely its job -- so requiring
-    them here would misdescribe what it accepts.
-    """
-
-    description: str
-    unit_price_gross: Decimal
-    discount_amount: Decimal
-    net_amount: Decimal
-    vat_amount: Decimal
-    vat_rate: Decimal
-
-
-class DerivedTotals(t.TypedDict):
-    """The header money columns as computed from an invoice's line items.
-
-    Keys mirror :data:`DERIVED_TOTAL_FIELDS` one-for-one; keep the two in sync.
-    """
-
-    total_gross: Decimal
-    total_net: Decimal
-    total_vat: Decimal
-    discount_amount_total: Decimal
-    vat_rate: Decimal
-
-
-# Optional string columns (blank=True, default="") whose schema fields are nullable:
-# the FE sends `null` for a cleared optional field, so coerce None -> "" to respect
-# the NOT NULL DB constraint. Only fields that are genuinely blankable belong here —
-# required columns (e.g. buyer_name, currency) lack blank=True, so an empty string is
-# not a valid value for them. This set doubles as the exemption list for the null guard
-# in update_draft_invoice (#908): a null on any field outside it is rejected with a 422
-# naming that field, before full_clean() is ever reached.
-_BLANKABLE_STRING_FIELDS = frozenset(
-    {
-        "buyer_vat_id",
-        "buyer_vat_country",
-        "buyer_address",
-        "buyer_email",
-        "discount_code_text",
-    }
-)
-
-
-def _normalize_line_items(line_items: list[SubmittedLineItem]) -> list[InvoiceLineItemDict]:
-    """Coerce submitted line items to the stored shape: every key present, every amount a string.
-
-    ``InvoiceLineItemSchema`` parses the amounts as ``Decimal``; the JSON column
-    stores them as strings, and :func:`_derive_totals` and
-    :attr:`AttendeeInvoice.vat_breakdown` both read them back that way.
-    """
-    return [
-        InvoiceLineItemDict(
-            description=item.get("description", ""),
-            unit_price_gross=str(item.get("unit_price_gross", "0.00")),
-            discount_amount=str(item.get("discount_amount", "0.00")),
-            net_amount=str(item.get("net_amount", "0.00")),
-            vat_amount=str(item.get("vat_amount", "0.00")),
-            vat_rate=str(item.get("vat_rate", "0.00")),
-        )
-        for item in line_items
-    ]
-
-
-def _derive_totals(line_items: list[InvoiceLineItemDict]) -> DerivedTotals:
-    """Compute an invoice's header money columns from its line items.
-
-    Mirrors :func:`generate_attendee_invoice` exactly: that function sums the very
-    per-payment amounts :func:`_build_line_items` writes into each line, so a
-    recomputed header is identical to a freshly generated one -- no rounding, no
-    tolerance, both sides are plain sums of the same stored decimals.
-
-    ``vat_rate`` follows generation's "first payment's rate" convention and is
-    therefore just as lossy on a mixed-rate cart; consult
-    :attr:`AttendeeInvoice.has_mixed_vat_rates` before rendering it.
-
-    Reads the four keys :attr:`AttendeeInvoice.vat_breakdown` reads directly, so
-    a row malformed enough to break one already breaks the other. ``discount_amount``
-    is the exception and is read defensively: it is the one key the breakdown never
-    touches, and this function runs inside the issuance guard on rows nobody
-    normalized, where a missing key would be a 500 on a document the guard exists
-    to judge. Absent means no discount, which is what every writer stores anyway.
-    """
-    zero = Decimal("0.00")
-    return DerivedTotals(
-        total_gross=sum((Decimal(item["unit_price_gross"]) for item in line_items), zero),
-        total_net=sum((Decimal(item["net_amount"]) for item in line_items), zero),
-        total_vat=sum((Decimal(item["vat_amount"]) for item in line_items), zero),
-        discount_amount_total=sum((Decimal(item.get("discount_amount", "0.00")) for item in line_items), zero),
-        vat_rate=Decimal(line_items[0]["vat_rate"]) if line_items else zero,
-    )
-
-
-def _assert_totals_reconcile(invoice: AttendeeInvoice) -> None:
-    """Refuse to issue an invoice whose header totals disagree with its line items.
-
-    :func:`update_draft_invoice` derives the header from the lines, so a draft
-    edited through the API always reconciles. This guards the other door (#911):
-    a row that drifted before that landed, or one desynced by a raw
-    ``QuerySet.update()`` or a shell write that bypasses the service layer, must
-    not become a legal document and a delivered PDF on the strength of a header
-    its own ``vat_breakdown`` contradicts. The Django admin is *not* such a door
-    -- ``AttendeeInvoiceAdmin`` lists every money column and ``line_items`` as
-    read-only -- so do not relax that admin without revisiting this.
-
-    Recovery for such a row is a PATCH of ``line_items``, which recomputes the
-    header from whatever the organizer states the lines to be.
-
-    Only the three figures a buyer is legally billed on are checked. ``vat_rate``
-    is excluded on purpose: it is a documented lossy scalar (the first line's
-    rate, meaningless on a mixed-rate cart, which is what ``has_mixed_vat_rates``
-    is for), and refusing to issue a correct invoice over a label would be
-    disproportionate. ``discount_amount_total`` is likewise informational.
-
-    Raises:
-        HttpError 422: If any reconciled total differs from its line-item sum.
-    """
-    derived = _derive_totals(invoice.line_items)
-    mismatched = [
-        # Symbolic, not prose: the message around it is translated, this is not.
-        f"{name} ({actual} != {expected})"
-        for name, actual, expected in (
-            ("total_gross", invoice.total_gross, derived["total_gross"]),
-            ("total_net", invoice.total_net, derived["total_net"]),
-            ("total_vat", invoice.total_vat, derived["total_vat"]),
-        )
-        if actual != expected
-    ]
-    if mismatched:
-        raise HttpError(
-            422,
-            str(_("Invoice totals do not reconcile with its line items: {details}")).format(
-                details="; ".join(mismatched)
-            ),
-        )
-
-
-def _apply_draft_update(invoice: AttendeeInvoice, update_data: dict[str, t.Any]) -> AttendeeInvoice:
-    """Write an already-validated payload onto the invoice, under a row lock.
-
-    Re-reads the row with ``select_for_update`` and re-checks DRAFT before writing.
-    The caller's instance was fetched by the controller before any validation ran,
-    so its status is a snapshot: a concurrent issue landing in that window would
-    leave an unlocked write editing an already-issued legal document (#911 review).
-
-    Split out from :func:`update_draft_invoice` so the lock covers the write and
-    nothing else -- rejecting a malformed payload should not take a row lock.
-
-    Returns:
-        The locked, updated invoice -- a different instance from the one passed in.
-
-    Raises:
-        HttpError 409: If the locked row is no longer a draft.
-    """
-    with transaction.atomic():
-        invoice = AttendeeInvoice.objects.select_for_update().get(pk=invoice.pk)
-        if invoice.status != AttendeeInvoice.InvoiceStatus.DRAFT:
-            raise HttpError(409, str(_("Only draft invoices can be edited.")))
-
-        for field, value in update_data.items():
-            setattr(invoice, field, value)
-
-        # Invalidate stale PDF so it gets regenerated on next download
-        if invoice.pdf_file:
-            invoice.pdf_file.delete(save=False)
-            update_data["pdf_file"] = ""
-
-        invoice.save(update_fields=[*update_data.keys(), "updated_at"])
-        return invoice
-
-
-def update_draft_invoice(
-    invoice: AttendeeInvoice,
-    update_data: dict[str, t.Any],
-) -> AttendeeInvoice:
-    """Update a DRAFT invoice. Only drafts are editable.
-
-    Editing ``line_items`` recomputes every column in :data:`DERIVED_TOTAL_FIELDS`
-    from the new lines (#911). Those columns are not in
-    :data:`EDITABLE_DRAFT_FIELDS`, so an organizer who wants a different total
-    states it as line items -- which is what an invoice is -- and the header can
-    never contradict the ``vat_breakdown`` summed from the same rows.
-
-    Args:
-        invoice: The invoice to update.
-        update_data: Dict of fields to update (from exclude_unset).
-
-    The write itself happens in :func:`_apply_draft_update`, on a row-locked
-    re-read, so the returned instance is not the one passed in. The status is
-    checked twice on purpose: once here to fail fast on the caller's instance,
-    once under the lock because only that row is current.
-
-    Returns:
-        The updated invoice, re-read under the lock.
-
-    Raises:
-        HttpError 409: If the invoice is not a draft, either as fetched or by the
-            time the row lock is acquired.
-        HttpError 422: If update_data contains disallowed fields (including a
-            derived total), or an explicit null for a field that is not one of
-            the blankable optional strings.
-    """
-    if invoice.status != AttendeeInvoice.InvoiceStatus.DRAFT:
-        raise HttpError(409, str(_("Only draft invoices can be edited.")))
-
-    if not update_data:
-        return invoice
-
-    disallowed = set(update_data.keys()) - EDITABLE_DRAFT_FIELDS
-    if disallowed:
-        raise HttpError(422, str(_("Cannot edit fields: {fields}")).format(fields=", ".join(sorted(disallowed))))
-
-    nulled = {field for field, value in update_data.items() if value is None} - _BLANKABLE_STRING_FIELDS
-    if nulled:
-        raise HttpError(422, str(_("Fields cannot be null: {fields}")).format(fields=", ".join(sorted(nulled))))
-
-    if "line_items" in update_data:
-        update_data["line_items"] = _normalize_line_items(update_data["line_items"])
-        # The header follows the lines it summarizes and is never stated alongside
-        # them (#911). Recomputed only on an edit that touches the lines: doing it
-        # unconditionally would zero the totals of any row whose line_items happen
-        # to be empty.
-        update_data.update(_derive_totals(update_data["line_items"]))
-
-    # Coerce None -> "" for blankable string columns (FE sends null for cleared fields).
-    for field in _BLANKABLE_STRING_FIELDS:
-        if field in update_data and update_data[field] is None:
-            update_data[field] = ""
-
-    return _apply_draft_update(invoice, update_data)
-
-
-def issue_draft_invoice(invoice: AttendeeInvoice) -> AttendeeInvoice:
-    """Issue a DRAFT invoice: finalize, set issued_at, regenerate PDF.
-
-    Idempotent: if the invoice is already ISSUED (e.g., PDF generation failed
-    on a previous attempt), re-generates the PDF and re-delivers.
-
-    Args:
-        invoice: The draft or already-issued invoice.
-
-    Both the status check and the reconciliation run on a row-locked re-read, so
-    the returned instance is not the one passed in.
-
-    Returns:
-        The issued invoice, re-read under the lock.
-
-    Raises:
-        HttpError 409: If the invoice is CANCELLED.
-        HttpError 422: If a DRAFT's header totals do not reconcile with its line
-            items -- such an invoice must not become a legal document (#911).
-    """
-    with transaction.atomic():
-        # Lock and re-read: the reconciliation below decides whether this row may
-        # become a legal document, so it must judge the committed row, not the
-        # snapshot the controller fetched. Unlocked, a concurrent PATCH landing
-        # after that fetch would be validated in absentia and then rendered into
-        # a PDF that contradicts what the database actually holds.
-        invoice = AttendeeInvoice.objects.select_for_update().get(pk=invoice.pk)
-
-        if invoice.status == AttendeeInvoice.InvoiceStatus.CANCELLED:
-            raise HttpError(409, str(_("Cancelled invoices cannot be issued.")))
-
-        if invoice.status == AttendeeInvoice.InvoiceStatus.DRAFT:
-            # Only on the DRAFT -> ISSUED transition: re-issuing an already-ISSUED
-            # invoice exists to recover a failed PDF generation, and that recovery
-            # must stay available even for a row that drifted before #911 landed.
-            _assert_totals_reconcile(invoice)
-            invoice.status = AttendeeInvoice.InvoiceStatus.ISSUED
-            invoice.issued_at = timezone.now()
-            invoice.save(update_fields=["status", "issued_at", "updated_at"])
-
-    # (Re-)generate PDF without DRAFT watermark. Outside the lock: WeasyPrint is
-    # slow, and once the status is committed no edit can land -- update_draft_invoice
-    # refuses a non-DRAFT row under that same lock.
-    _generate_and_save_pdf(invoice)
-
-    return invoice
-
-
-def issue_and_deliver(invoice: AttendeeInvoice) -> AttendeeInvoice:
-    """Issue a DRAFT invoice and enqueue background delivery to the buyer.
-
-    Thin orchestration helper for the organization admin "issue" endpoint:
-    finalizes the invoice via :func:`issue_draft_invoice` and then schedules
-    the delivery task. Keeps celery dispatch out of the controller layer.
-
-    Args:
-        invoice: The draft or already-issued invoice.
-
-    Returns:
-        The issued invoice.
-    """
-    from events.tasks import deliver_attendee_invoice_task
-
-    invoice = issue_draft_invoice(invoice)
-    # Defer delivery until the issuance write commits so the worker reads
-    # the issued invoice. With ATOMIC_REQUESTS=True an immediate .delay()
-    # would race the request commit.
-    transaction.on_commit(lambda: deliver_attendee_invoice_task.delay(str(invoice.id)))
-    return invoice
-
-
-def delete_draft_invoice(invoice: AttendeeInvoice) -> None:
-    """Delete a DRAFT invoice.
-
-    Raises:
-        HttpError 409: If the invoice is not a draft.
-    """
-    if invoice.status != AttendeeInvoice.InvoiceStatus.DRAFT:
-        raise HttpError(409, str(_("Only draft invoices can be deleted.")))
-
-    # Delete PDF file if it exists
-    if invoice.pdf_file:
-        invoice.pdf_file.delete(save=False)
-
-    invoice.delete()
-
-
-# ---------------------------------------------------------------------------
 # Phase 5: Email delivery
 # ---------------------------------------------------------------------------
 
@@ -912,6 +559,11 @@ def generate_attendee_credit_note(
 
     # If invoice is still a draft, just delete it instead of creating a credit note
     if invoice.status == AttendeeInvoice.InvoiceStatus.DRAFT:
+        # Imported here, not at module scope: the draft module imports
+        # _generate_and_save_pdf from this one, so a top-level import would close
+        # the cycle. Same pattern as issue_and_deliver's task import.
+        from events.service.attendee_invoice_draft_service import delete_draft_invoice
+
         delete_draft_invoice(invoice)
         return None
