+++ src/events/tests/test_attendee_invoice_service.py
@@ -24,14 +24,16 @@ from accounts.models import RevelUser
 from events.models import Event, Organization, Ticket, TicketTier
 from events.models.attendee_invoice import AttendeeInvoice
 from events.models.ticket import Payment
-from events.service.attendee_invoice_service import (
+from events.service.attendee_invoice_draft_service import (
     DERIVED_TOTAL_FIELDS,
-    _is_export,
     delete_draft_invoice,
-    generate_attendee_invoice,
     issue_draft_invoice,
-    set_invoicing_mode,
     update_draft_invoice,
+)
+from events.service.attendee_invoice_service import (
+    _is_export,
+    generate_attendee_invoice,
+    set_invoicing_mode,
     validate_invoicing_prerequisites,
 )
